+++ src/hexsample/tasks.py
@@ -167,8 +167,7 @@ class ReconstructionDefaults:
     eta_2pix_rad: float = 0.127
     eta_3pix_rad0: float = 0.513
     eta_3pix_rad1: float = 0.141
-    eta_3pix_theta0: float = 0.074
-    eta_3pix_theta1: float = -0.5
+    eta_3pix_theta0: float = 0.104
 
 
 def reconstruct(
@@ -182,7 +181,6 @@ def reconstruct(
         eta_3pix_rad0: float = ReconstructionDefaults.eta_3pix_rad0,
         eta_3pix_rad1: float = ReconstructionDefaults.eta_3pix_rad1,
         eta_3pix_theta0: float = ReconstructionDefaults.eta_3pix_theta0,
-        eta_3pix_theta1: float = ReconstructionDefaults.eta_3pix_theta1,
         header_kwargs: dict = None,
         ) -> str:
     """Run the reconstruction.
@@ -256,7 +254,7 @@ def reconstruct(
             args = event.trigger_id, event.timestamp(), event.livetime, cluster
             recon_event = ReconEvent(*args, pos_recon_algorithm, readout.pitch,
                                      eta_2pix_rad, eta_3pix_rad0, eta_3pix_rad1,
-                                     eta_3pix_theta0, eta_3pix_theta1)
+                                     eta_3pix_theta0)
             mc_event = input_file.mc_event(i)
             output_file.add_row(recon_event, mc_event)
     output_file.flush()